Move common methods of mainchain/sidechain to BaseInteroperabilityMethod
